auth in generate_tournament_seats.py

+++ generate_tournament_seats.py
@@ -2,31 +2,28 @@ import json
 import os
 
 import requests
+import random
 
 from crawler import get_bearer_token
 
+POLEMICA_USERNAME = os.getenv("POLEMICA_USERNAME")
+POLEMICA_PASSWORD = os.getenv("POLEMICA_PASSWORD")
+
+random.seed(42)
 token = get_bearer_token()
 headers = {'Authorization': f'Bearer {token}'}
 baseurl = "https://app.polemicagame.com/v1"
-competition_id = "2695"
+competition_id = "2873"
 members_url = baseurl + "/competitions/" + competition_id + "/members"
 games_url = baseurl + "/competitions/" + competition_id + "/games"
 tour_games_dir = "tour_games"
 
-seats = {1: {1: {1: 0, 2: 1, 3: 2, 4: 12, 5: 10, 6: 5, 7: 16, 8: 17, 9: 19, 10: 14},
-             2: {1: 4, 2: 3, 3: 8, 4: 13, 5: 15, 6: 6, 7: 18, 8: 7, 9: 9, 10: 11}},
-         2: {1: {1: 2, 2: 15, 3: 19, 4: 9, 5: 3, 6: 0, 7: 6, 8: 14, 9: 16, 10: 12},
-             2: {1: 8, 2: 17, 3: 18, 4: 4, 5: 7, 6: 11, 7: 1, 8: 13, 9: 5, 10: 10}},
-         3: {1: {1: 13, 2: 8, 3: 5, 4: 1, 5: 6, 6: 14, 7: 9, 8: 16, 9: 18, 10: 0},
-             2: {1: 19, 2: 10, 3: 12, 4: 17, 5: 11, 6: 4, 7: 7, 8: 3, 9: 2, 10: 15}},
-         4: {1: {1: 18, 2: 12, 3: 3, 4: 5, 5: 14, 6: 19, 7: 17, 8: 6, 9: 13, 10: 7},
-             2: {1: 9, 2: 2, 3: 16, 4: 10, 5: 4, 6: 8, 7: 11, 8: 15, 9: 0, 10: 1}},
-         5: {1: {1: 15, 2: 16, 3: 14, 4: 6, 5: 5, 6: 10, 7: 3, 8: 4, 9: 1, 10: 19},
-             2: {1: 7, 2: 9, 3: 17, 4: 11, 5: 2, 6: 12, 7: 0, 8: 18, 9: 8, 10: 13}},
-         6: {1: {1: 17, 2: 13, 3: 4, 4: 8, 5: 19, 6: 16, 7: 14, 8: 11, 9: 7, 10: 9},
-             2: {1: 12, 2: 5, 3: 0, 4: 15, 5: 1, 6: 18, 7: 2, 8: 10, 9: 3, 10: 6}},
-         7: {1: {1: 16, 2: 19, 3: 7, 4: 0, 5: 12, 6: 3, 7: 8, 8: 1, 9: 4, 10: 18},
-             2: {1: 5, 2: 11, 3: 15, 4: 14, 5: 17, 6: 13, 7: 10, 8: 9, 9: 6, 10: 2}}}
+
+def authorize():
+    global token, headers
+    token = requests.post(baseurl + "/auth/login",
+                          data={"username": POLEMICA_USERNAME, "password": POLEMICA_PASSWORD}).json()["access_token"]
+    headers = {'Authorization': f'Bearer {token}'}
 
 
 def crawl_games():
@@ -41,7 +38,7 @@ def crawl_games():
 
 def print_members():
     members = requests.get(members_url, headers=headers).json()
-    print(list(map(lambda it: it['player']['id'], members)))
+    # print(list(map(lambda it: it['player']['id'], members)))
     for member in members:
         print(member['player']['id'], member['player']['username'])
 
@@ -49,12 +46,12 @@ def print_members():
 def delete_games():
     games = requests.get(games_url, headers=headers).json()
     for game_id in games:
-        requests.delete(games_url + "/" + str(game_id["id"]), headers=headers)
+        print("Delete", requests.delete(games_url + "/" + str(game_id["id"]), headers=headers))
 
 
 def save_modified_games():
     members = requests.get(members_url, headers=headers).json()
-
+    # random.shuffle(masters)
     for filename in os.listdir(tour_games_dir):
         if not filename.endswith('.json'):
             continue
@@ -65,18 +62,23 @@ def save_modified_games():
             game_num = game["num"]
             table = game["table"]
             game["id"] = None
-
-            # for player in game["players"]:
-            #     member = members[seats[game_num][table][player["position"]] - 1]
+            game["isLive"] = True
+            # master = masters[game_num - 1]
+            # gameMembers = list(filter(lambda it: it["player"]["id"] != master, members))
+            # random.shuffle(gameMembers)
+            # for player, member in zip(game["players"], gameMembers):
             #     player["player"]["id"] = member["player"]["id"]
             #     player["username"] = member["player"]["username"]
-            print(game["players"])
-            print(requests.post(games_url, json=game, headers=headers))
+            # game["master"] = master
+            # game["referee"]["id"] = master
+            res = requests.post(games_url, json=game, headers=headers)
+            print(res.status_code, res.text)
 
 
 def save_modified_games_in_overlay_service():
     # Base URLs for your API
     host = 'http://51.250.18.236:8090'  # Replace with your actual host
+    # host = 'http://localhost:8080'  # Replace with your actual host
     overlay_game_players_url = f'{host}/gamePlayers'
     overlay_games_url = f'{host}/games'
     headers = {
@@ -98,6 +100,7 @@ def save_modified_games_in_overlay_service():
                 player_data = {
                     "nickname": player["username"],
                     "checks": [],  # Adjust this if you have actual check data
+                    "guess": [],  # Adjust this if you have actual check data
                     "stat": {},  # Adjust this if you have actual stat data
                     "role": player.get("role", "red"),  # Default role to "red" if not provided
                     "place": player["position"],
@@ -116,8 +119,8 @@ def save_modified_games_in_overlay_service():
 
             # After all players are created, create the game
             game_data = {
-                "type": "CUSTOM",
-                    "tournamentId": 2695,  # Replace with your actual tournament ID
+                "type": "POLEMICA",
+                "tournamentId": competition_id,  # Replace with your actual tournament ID
                 "tableNum": table,
                 "gameNum": game_num,
                 "players": players_urls
@@ -134,8 +137,9 @@ def save_modified_games_in_overlay_service():
 
 
 if __name__ == "__main__":
-    # crawl_games()
-    # print_members()
+    authorize()
+    crawl_games()
+    print_members()
     delete_games()
     save_modified_games()
-    # save_modified_games_in_overlay_service()
+    save_modified_games_in_overlay_service()
